Update the student.statements.rda and benchmark.rda package data objects

Imports the results of a Flex Query run within IBKR and downloaded as an xml file, fetches the relevant data, and saves a tibble containing the student statement information needed to calculate standings and performance metrics as the package data object student_reports.rda. Same for the benchmark.

refresh_all_data()

Arguments

filename

Name of the xml file containing student statement data.

Refreshes benchmark.rda, gspc.rda, performance_reports.rda, rfr.rda, and student_reports.rda. Stores as package data objects.